fix: inline p5.js for windows users
Microsoft Edge will run files in an extremely strict execution context.
It will not even load adjacent files from disk, so we need to inline
everything to get it to work. Of course, users who install node can just
do the live-reloading server and things are way smoother, but we want to
maintain an extremely low barrier to entry here.